requirements.txt
git+https://github.com/chrisglass/xhtml2pdf.git
django-easy-pdf
